Add files via upload
Fixed an error within Find that caused it to crash if returning the first item of a list
Got Erase working for all cases as far as my testing goes. For now I'll be focusing a lot more on freecodecamp.org and project crab hammer but before long I'll have a String class and a double linked list for this repository as well.
post: My testing methods throughout this project were somewhat weak forcing me to go back and redo some methods heavily (or from scratch in the case of Erase) so in the future building better test cases is something I'll be looking to focus on.